+++ docs/conf.py
@@ -3,13 +3,18 @@
 # For the full list of built-in configuration values, see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
+import os
+import sys
+sys.path.insert(0, os.path.abspath('..'))
+
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'Highway Circuit Breaker'
-copyright = '2025, Your Organization'
-author = 'Your Name'
+project = 'Resilient Circuit'
+copyright = '2025, Resilient Circuit Contributors'
+author = 'Resilient Circuit Team'
 release = '0.3.0'
+version = '0.3.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -19,16 +24,45 @@ extensions = [
     'sphinx.ext.viewcode',
     'sphinx.ext.napoleon',
     'sphinx.ext.intersphinx',
+    'sphinx.ext.autosummary',
+    'sphinx.ext.coverage',
+    'sphinx.ext.todo',
+    'sphinx_rtd_theme',
 ]
 
 templates_path = ['_templates']
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+
+# The suffix of source filenames
+source_suffix = {
+    '.rst': 'restructuredtext',
+}
+
+# The master toctree document
+master_doc = 'index'
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
+
+# Theme options
+html_theme_options = {
+    'display_version': True,
+    'prev_next_buttons_location': 'bottom',
+    'style_external_links': False,
+    'navigation_depth': 4,
+    # Toc options
+    'collapse_navigation': False,
+    'sticky_navigation': True,
+    'includehidden': True,
+    'titles_only': False
+}
+
+# Add logo and favicon (if you have them)
+# html_logo = '_static/logo.png'
+# html_favicon = '_static/favicon.ico'
 
 # -- Extension configuration -------------------------------------------------
 
@@ -41,8 +75,37 @@ autodoc_default_options = {
     'exclude-members': '__weakref__'
 }
 
+autodoc_typehints = 'description'
+autodoc_class_signature = 'separated'
+
+# Napoleon settings for Google and NumPy style docstrings
+napoleon_google_docstring = True
+napoleon_numpy_docstring = True
+napoleon_include_init_with_doc = False
+napoleon_include_private_with_doc = False
+napoleon_include_special_with_doc = True
+napoleon_use_admonition_for_examples = False
+napoleon_use_admonition_for_notes = False
+napoleon_use_admonition_for_references = False
+napoleon_use_ivar = False
+napoleon_use_param = True
+napoleon_use_rtype = True
+napoleon_preprocess_types = False
+napoleon_type_aliases = None
+napoleon_attr_annotations = True
+
 # Intersphinx mapping
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
-    'requests': ('https://requests.readthedocs.io/en/latest/', None),
+    'psycopg': ('https://www.psycopg.org/psycopg3/docs/', None),
 }
+
+# Autosummary settings
+autosummary_generate = True
+
+# Todo extension settings
+todo_include_todos = True
+
+# Add any custom CSS or JS files
+# html_css_files = []
+# html_js_files = []